backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
python-multipart
numpy
opencv-python-headless
ImageHash
Pillow
